requirements.txt
streamlit==1.56.0
mysql-connector-python==9.6.0
pandas==3.0.2
openpyxl==3.1.5
